# docs/openapi3.yaml
openapi: 3.0.0
...
info:
  contact:
    email: shpakovskiipa@gmail.com
    name: API Support
  description: This is a Eschool API.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Eschool API
  version: "1.0"
paths:
...
  /courses:
    get:
      description: get all courses
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                items:
                  $ref: "#/components/schemas/github_com_paw1a_eschool_internal_adapter_delivery_\
                    http_v1_dto.CourseDTO"
                type: array
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/internal_adapter_delivery_http_v1.RestErrorInternal\
                  Error"
      summary: GetAllCourses
      tags:
        - course
...
  "/courses/{id}":
    get:
      description: get course by id
      parameters:
        - description: course id
          in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                items:
                  $ref: "#/components/schemas/github_com_paw1a_eschool_internal_adapter_delivery_\
                    http_v1_dto.CourseDTO"
                type: array
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/internal_adapter_delivery_http_v1.RestErrorBadReque\
                  st"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/internal_adapter_delivery_http_v1.RestErrorNotFound"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/internal_adapter_delivery_http_v1.RestErrorInternal\
                  Error"
      summary: GetCourseByID
      tags:
        - course
...
  /users:
    get:
      description: get all users
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                items:
                  $ref: "#/components/schemas/github_com_paw1a_eschool_internal_adapter_delivery_\
                    http_v1_dto.UserDTO"
                type: array
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/internal_adapter_delivery_http_v1.RestErrorInternal\
                  Error"
      summary: GetAllUsers
      tags:
        - user
    put:
      description: update user information
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/github_com_paw1a_eschool_internal_adapter_delivery_\
                http_v1_dto.UpdateUserDTO"
        description: update user info
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                items:
                  $ref: "#/components/schemas/github_com_paw1a_eschool_internal_adapter_delivery_\
                    http_v1_dto.UserDTO"
                type: array
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/internal_adapter_delivery_http_v1.RestErrorBadReque\
                  st"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/internal_adapter_delivery_http_v1.RestErrorUnauthor\
                  ized"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/internal_adapter_delivery_http_v1.RestErrorInternal\
                  Error"
      security:
        - ApiKeyAuth: []
      summary: UpdateUser
      tags:
        - user
  "/users/{id}":
    get:
      description: get user by id
      parameters:
        - description: user id
          in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                items:
                  $ref: "#/components/schemas/github_com_paw1a_eschool_internal_adapter_delivery_\
                    http_v1_dto.UserDTO"
                type: array
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/internal_adapter_delivery_http_v1.RestErrorBadReque\
                  st"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/internal_adapter_delivery_http_v1.RestErrorNotFound"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/internal_adapter_delivery_http_v1.RestErrorInternal\
                  Error"
      summary: GetUserByID
      tags:
        - user
  /users/account:
    get:
      description: get user account
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                items:
                  $ref: "#/components/schemas/github_com_paw1a_eschool_internal_adapter_delivery_\
                    http_v1_dto.UserInfoDTO"
                type: array
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/internal_adapter_delivery_http_v1.RestErrorUnauthor\
                  ized"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/internal_adapter_delivery_http_v1.RestErrorInternal\
                  Error"
      security:
        - ApiKeyAuth: []
      summary: GetUserAccount
      tags:
        - user
  /users/courses:
    get:
      description: find user courses
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                items:
                  $ref: "#/components/schemas/github_com_paw1a_eschool_internal_adapter_delivery_\
                    http_v1_dto.CourseDTO"
                type: array
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/internal_adapter_delivery_http_v1.RestErrorBadReque\
                  st"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/internal_adapter_delivery_http_v1.RestErrorUnauthor\
                  ized"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/internal_adapter_delivery_http_v1.RestErrorInternal\
                  Error"
      security:
        - ApiKeyAuth: []
      summary: FindUserCourses
      tags:
        - user
  "/users/courses/{id}":
    put:
      description: add user free course
      parameters:
        - description: course id
          in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: string
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/internal_adapter_delivery_http_v1.RestErrorBadReque\
                  st"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/internal_adapter_delivery_http_v1.RestErrorUnauthor\
                  ized"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/internal_adapter_delivery_http_v1.RestErrorForbidde\
                  n"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/internal_adapter_delivery_http_v1.RestErrorInternal\
                  Error"
      security:
        - ApiKeyAuth: []
      summary: AddUserFreeCourse
      tags:
        - user
servers:
  - url: //localhost:8080/api/v1
components:
  securitySchemes:
    ApiKeyAuth:
      in: header
      name: Authorization
      type: apiKey
  schemas:
...
    internal_adapter_delivery_http_v1.RestErrorNotFound:
      properties:
        error:
          example: not found
          type: string
        status:
          example: 404
          type: integer
        timestamp:
          example: 2020-11-10T23:00:00+00:00
          type: string
      type: object
...